refactor(web): app route 수정
