+++ apps/web/src/app/[space-slug]/layout.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from "react";
 import { SpaceSidebar } from "@/_pages/space";
 import { getSpaceContext } from "@/features/space/lib/get-space-context";
+import { QueryProvider } from "@/shared/providers/query-client-provider";
 
 // layout에서 인증·스페이스·멤버십 검증을 수행한다.
 // getSpaceBySlug, getSpaceMembership에 React.cache()가 적용되어 있으므로
@@ -13,11 +14,14 @@ export default async function SpaceLayout({
   params: Promise<{ "space-slug": string }>;
 }) {
   const slug = (await params)["space-slug"];
+
   const { space, membership } = await getSpaceContext(slug);
 
   return (
-    <SpaceSidebar space={space} membership={membership}>
-      <main className="mx-auto w-full max-w-6xl p-6">{children}</main>
-    </SpaceSidebar>
+    <QueryProvider>
+      <SpaceSidebar space={space} membership={membership}>
+        <main className="mx-auto w-full max-w-6xl p-6">{children}</main>
+      </SpaceSidebar>
+    </QueryProvider>
   );
 }
