+++ apps/web/src/app/layout.tsx
@@ -1,10 +1,6 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-
-import { QueryProvider } from "@/shared/providers/query-client-provider";
 import "./globals.css";
-import { Toaster } from "@moum-zip/ui/components";
-import { NavigationMenu } from "@/shared/ui/navigation-menu";
 
 const pretendard = localFont({
   src: "./fonts/PretendardVariable.woff2",
@@ -24,12 +20,8 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <body className={`${pretendard.className} ${pretendard.variable} antialiased`}>
-        <QueryProvider>
-          <NavigationMenu />
-          <main>{children}</main>
-          <Toaster />
-        </QueryProvider>
+      <body className={`${pretendard.className} ${pretendard.variable} bg-background-secondary antialiased`}>
+        {children}
       </body>
     </html>
   );
